+++ R/00_global_variables.R
@@ -6,6 +6,6 @@ globalVariables(
     "cmy","lambda","location","m","max_est","mean_log","method","min_est",
     "name","prob","rate","sd_log","shape1","shape2","size","total","total_deviance",
     "abs_aic","aic_value","data","dist_with_params","ks","lm","lm_model","mu",
-    "stan_dev","tidy_ks","prop","dens_tbl"
+    "stan_dev","tidy_ks","prop","dens_tbl","cih","cil","mstat","stat"
   )
 )